added redux

+++ src/app/admin/product-form/product-form.component.ts
@@ -5,6 +5,10 @@ import { Router, ActivatedRoute } from '@angular/router';
 import 'rxjs/add/operator/take';
 import { FormGroup, FormControl, Validators } from '@angular/forms';
 import { PriceValidators } from './price.validators';
+import { NgRedux, select } from 'ng2-redux';
+import { IAppState } from '../../store';
+import { INCREMENT } from '../../actions';
+import { Map } from 'immutable';
 
 
 
@@ -14,6 +18,11 @@ import { PriceValidators } from './price.validators';
   styleUrls: ['./product-form.component.css']
 })
 export class ProductFormComponent implements OnInit {
+  //REDUX VERSION 2
+  //@select(s => s.get('counter')) count;
+
+  @select('counter') count;
+ 
   categories$;
   product = {};
   id;
@@ -26,14 +35,20 @@ export class ProductFormComponent implements OnInit {
   });
 
   constructor(
+    // REDUX VERSION 2
+    //private ngRedux: NgRedux<Map<string, any>>,
+    private ngRedux: NgRedux<IAppState>,
     private route: ActivatedRoute,
     private categoryService: CategoryService, 
     private productService: ProductService, 
     private router: Router) {
     this.categories$ = categoryService.getAll();
+    
 
     this.id = this.route.snapshot.paramMap.get('id');
     if (this.id) this.productService.get(this.id).take(1).subscribe(p => this.product = p);
+
+    
    }
 
    save(){
@@ -44,11 +59,15 @@ export class ProductFormComponent implements OnInit {
    }
 
    delete(){
-     //if (!confirm('Are you sure you want to delete this product?')) return;
+     if (!confirm('Are you sure you want to delete this product?')) return;
      
        this.productService.delete(this.id);
        this.router.navigate(['/admin/products']);
      
+   }
+
+   increment(){
+     this.ngRedux.dispatch({ type: INCREMENT });
    }
 
   ngOnInit() {
